Topnav bar, converted buttons to images, text if no requests to display
Added a topnav bar with "DocDocGo", currently doesn't do anything other
than look nice. Converted the ambulance and docdocgo buttons to
clickable images. Added image in the topNav bar as well that will be
changed to the DocDocGo logo for the final version. If there are no
requests to display, a text label be will displayed saying "No Requests
to Display"